+++ src/front/js/layout.js
@@ -26,7 +26,6 @@ import { Donate } from "./pages/donate";
 import { ToastContainer } from "react-toastify";
 
 import "react-toastify/dist/ReactToastify.css";
-import { Navbar } from "./component/navbar";
 
 //create your first component
 const Layout = () => {
@@ -45,7 +44,7 @@ const Layout = () => {
 				{location.pathname !== "/login" &&
 				location.pathname !== "/signup" &&
 				location.pathname !== "/reset_pass" ? (
-					<Navbar />
+					<Header />
 				) : null}
 				<ToastContainer />
 				<Switch>
