--- src/front/styles/header2.scss
@@ -1,44 +1,0 @@
-
-
-.header-general{
-        background-color: #ffffff;
-        color: #7b7b7b;
-
-        .nav-item {
-             a.nav-link:link {
-                color: #7b7b7b;
-            }
-
-            a.nav-link:visited {
-               color: #7b7b7b;
-            }
-            
-            /* mouse over link */
-            a.nav-link:hover {
-                color: #08cc7f;
-            }
-
-            /* selected link */
-            a.nav-link:active {
-                color: #08cc7f;
-            }
-
-        }
-         
-        .custom-toggler .navbar-toggler-icon {
-  background-image: url("data:image/svg+xml;charset=utf8,%3Csvg viewBox='0 0 32 32' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath stroke='rgba(255,102,203, 0.5)' stroke-width='2' stroke-linecap='round' stroke-miterlimit='10' d='M4 8h24M4 16h24M4 24h24'/%3E%3C/svg%3E");
-}
-
-.custom-toggler.navbar-toggler {
-  border-color: rgb(255,102,203);
-} 
-        
-    
-    .bg-green {
-        color: #ffffff;
-        background-color: #08cc7f;
-    }
-
-    
-
-}
